\name{plmm-package}
\alias{plmm-package}
\docType{package}
\title{
Partially Linear Mixed Effects Model
}
\description{
This package fits the partially linear mixed effects model (semiparametric random intercept model) using kernel regression, without distributional assumptions for the random terms. Estimation procedure is an iterative generalized least squares type. A nonparametric heteroskedastic variance function is allowed for the regression error. Bootstrap resampling is provided for inference. The package implements bandwidth selection by an alternative cross validation for correlated data. 
}
\details{
\tabular{ll}{
Package: \tab plmm\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2012-11-28\cr
License: \tab GPL (>= 2)\cr
}


}
\author{
OHINATA Ren

Maintainer: OHINATA Ren <Ren.Ohinata@wiwi.uni-goettingen.de>
}
\references{
Ohinata R. (2012). Partially Linear Mixed Effects Model without Distributional Assumptions. Working Paper.
}
\keyword{ package }
\keyword{ regression }
\keyword{ nonparametric }

